requirements/test.txt
-r base.txt

# Code quality checks/Linting
# -----------------------------------------------------------------------------
black~=22.6.0
flake8-isort~=4.2.0
flake8~=5.0.4
isort~=5.10.1
mypy-extensions~=0.4.3
pyright~=1.1.265


# Stubs
# -----------------------------------------------------------------------------
django-stubs-ext~=0.5.0
django-stubs~=1.12.0


# Testing
# -----------------------------------------------------------------------------
factory-boy~=3.2.1
Faker~=13.15.1
pytest-cov~=3.0.0
pytest-django~=4.5.2
pytest-forked~=1.4.0
pytest-sugar~=0.9.5
pytest-xdist~=2.5.0
pytest~=7.1.2
tox~=3.25.1


# Coverage
# -----------------------------------------------------------------------------
coverage~=6.4.2
coveralls~=3.3.1
django-coverage-plugin~=2.0.3
